requirements.txt
aiohttp==3.12.15
fastapi==0.128.0
httpx==0.28.1
pydantic==2.12.5
python-dotenv==1.2.1
python-jose[cryptography]
python-multipart
uvicorn==0.40.0
email-validator
python-multipart